requirements.txt
pandas==1.4.2
pandas_gbq==0.17.4
protobuf==4.21.2
python_dateutil==2.8.2
google==3.0.0
google-analytics-data==0.11.2
google-api-core==2.7.1
google-api-python-client==2.43.0
google-auth==2.6.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
google-cloud==0.34.0
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.2.3
google-cloud-firestore==2.4.0
google-cloud-pubsub==2.12.0
google-cloud-secret-manager==2.10.0
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
